Moved packages that were missed in restructuring and updated test filter in gradle tasks.
